test: Refactor advanced cluster test cases to use CaptureOrMockTestCaseAndRun and delete unnecessary resource_unit_test.go
