+++ internal/service/advancedclustertpf/testdata/TestAccAdvancedCluster_shardedBasic.yaml
