+++ internal/service/advancedclustertpf/testdata/TestAccAdvancedCluster_shardedBasic/02_01_PATCH__api_atlas_v2_groups_{groupId}_clusters_{clusterName}_2024-08-05.json
