+++ internal/service/advancedclustertpf/testdata/TestAccClusterAdvancedClusterConfig_symmetricShardedOldSchema/01_01_POST__api_atlas_v2_groups_{groupId}_clusters_2024-10-23.json
