+++ internal/service/advancedclustertpf/testdata/TestAccClusterAdvancedClusterConfig_symmetricShardedOldSchema/02_01_PATCH__api_atlas_v2_groups_{groupId}_clusters_{clusterName}_2023-02-01.json
