+++ internal/service/advancedclustertpf/testdata/TestAccClusterAdvancedClusterConfig_symmetricShardedOldSchema/02_02_PATCH__api_atlas_v2_groups_{groupId}_clusters_{clusterName}_2024-10-23.json
