+++ internal/service/advancedclustertpf/testdata/TestAccClusterAdvancedClusterConfig_symmetricShardedOldSchemaDiskSizeGBAtElectableLevel.yaml
