+++ internal/service/advancedclustertpf/testdata/TestAccClusterAdvancedCluster_basicTenant.yaml
