+++ internal/service/advancedclustertpf/testdata/TestAccClusterAdvancedCluster_basicTenant/02_02_POST__api_atlas_v2_groups_{groupId}_clusters_2024-10-23.json
