+++ internal/service/advancedclustertpf/testdata/TestAccClusterAdvancedCluster_tenantUpgrade.yaml
